flash-intelligence-core/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
statsmodels==0.14.1
pyyaml==6.0.1
aiofiles==23.2.1
httpx==0.27.0
